Extract review grouping utils
